fix(aho-corasick): close the two quadratic build paths, and re-measure against the Regex baseline that does not allocate
Six review findings, all real.

The build had two paths quadratic in one state's branching factor, against a
public O(total pattern length) contract: threading the trie walked a state's
sibling list linearly to find an edge, and the flattening insertion-sorted each
row. A set of many one-character patterns reaches both — the UTF-16 alphabet
permits a root row tens of thousands of edges wide. The edge lookup is now one
hash probe per pattern character and the row sort is MemoryExtensions.Sort, and
a 4,000-single-character-pattern test pins it.

Scratch was sized from every supplied occurrence before duplicates collapsed, so
an array of a thousand references to one large string rented a thousand times
the buffers and could wrap the sum to a negative capacity. Duplicates are now
collapsed up front — which is also where the first-appearance ids come from, so
the trie no longer needs a duplicate check — and the remaining sum is checked.

CountMatches counted overlapping matches into an int that is not bounded by the
text length: a thousand nested patterns over a few million characters exceeds
int.MaxValue and silently returned a negative count. It returns long now; the
other tiers are bounded by an array or by the caller's buffer and stay int.

The Regex_FindAll baseline used the allocating MatchCollection when the net8.0
floor has Regex.EnumerateMatches over a span. That made the 2.4 MB allocation
column an artifact of the API chosen rather than of the structure — and the
allocation-free form is also 1.8x faster, so the ratio was wrong too. Both sides
now use their allocation-free enumerator, and every published figure is
re-derived: enumerate-all falls 11.4x -> 6.3x, membership 11.4x -> 11.1x, build
8.5x -> 4.7x (the hash-probe lookup costs ~9us), counting 3.2x -> 3.1x, and the
eight-pattern loss deepens to 9.2x. The allocation claim is gone from every
surface rather than restated.

And the XML docs claimed the Regex baseline wins the absent-pattern arm, which
contradicts every table in the PR. Removed.

+++ src/Celerity.Tests/Collections/AhoCorasickDifferentialTests.cs
@@ -90,7 +90,7 @@ public class AhoCorasickDifferentialTests
         PatternMatch[] actual = automaton.FindAll(text);
 
         Assert.Equal(expected, actual);
-        Assert.Equal(expected.Length, automaton.CountMatches(text));
+        Assert.Equal((long)expected.Length, automaton.CountMatches(text));
         Assert.Equal(expected.Length > 0, automaton.ContainsAny(text));
 
         Assert.Equal(expected.Length > 0, automaton.TryFindFirst(text, out PatternMatch first));
